HUE-141: Fixing bugs in the Hue Shell backend, simplifying code.
The Hue Shell app previously required opening subprocesses with shell=True, because the arguments passed in were
passed in as a long string, e.g. "pig -l /dev/null". Changing this to array format(["pig","-l","/dev/null"]) allows
us to use shell=False.  Also fixed a JS bug with the width of the input textarea, and simplified depender so that
instead of an absolute URL, which is vulnerable to port changes, it returns a relative URL, which is not.