HUE-141: Reducing changes to Tornado code for Hue Shell app.
Prior commits introduced yet another patch to Tornado that we would have to
track. This commit changes things so that is not necessary. Also had to move
around a couple of management commands to make this work.  Lastly, adding a
stream-based (i.e. offset based) reliability scheme as opposed to a chunk-based
one.